Load the python module based on the plugin name
pycrosswalk will look for a python module with the same as the plugin
(without the lib prefix and the .so extension on Linux) and at the same
directory as the plugin. PYXWALK_PATH is not needed anymore.

For instance, let's suppose the symbolic link bellow:

  /home/foo/libfoobar.so -> /usr/lib/pyxwalk/libpycrosswalk.so

The loader will try to load the following module:

  /home/foo/foobar.py

For the following command line:

  ./xwalk --external-extensions-path=/home/foo/