Support multiple python extensions
We trick Crosswalk to load multiple extensions using symbolic links, but
the reality is the extension only gets loaded once. The dynamic loader is smart
enough to figure that out.

So all the globals we have on the pycrosswalk plugin are shared. This is
fine, but we need to make sure that the next python extension being
loaded does not overwrite the callbacks set for the current extension.

The only tricky part was with the "instance_created/destroyed" callback,
which needed to be implemented using a closure library for C
(libffcall). I gave it for granted that closures were not possible in C,
but apparently it is using some stack manipulation magic.